refactor(tests): move data importer views tests
